requirements.txt
# Translation Studies Paper Collector
# Weekly collection of Chinese Translation Studies papers

flask==3.0.0
apscheduler==3.10.4
aiosmtplib==3.0.1
python-dotenv==1.0.0
feedparser==6.0.10
beautifulsoup4==4.12.2
requests==2.31.0
SQLAlchemy==2.0.23
